+++ dbt/macros/bcb.sql
@@ -1,0 +1,10 @@
+{% macro bcb_fonte_oficial() %}
+  https://dadosabertos.bcb.gov.br/
+{% endmacro %}
+
+{% macro bcb_sgs_columns() %}
+    sgs_codigo,
+    data,
+    valor,
+    ano
+{% endmacro %}